First attempt at one-off exiting non-zero
When running the m2ee cli program in one-off mode, like doing `m2ee
start`, the program will always exit cleanly, because errors are handled
and reported inside the cli loop. (see issue #5)

This change introduces a unchecked_onecmd on the cli that does not try
to cover up any of our own exception types, but logs the error and exits
1.

For this to actually work, it also means that errors which interrupt the
program flow actually need to be raised up until this place, which is
not always the case yet.