Remove runtime_blocking_connector option
Since 22160e16 we also had it disabled by default, because it appeared
to misbehave.

Newer Mendix versions will no longer have this option because it has no
meaning for newer Jetty versions.

As we don't want anyone to use it, just drop it. This option has always
been disabled by default in the Mendix Runtime, so defaults will do the
right thing for old versions.